Implement keyboard navigation
The principle is simple. `0123456789abcdefghijklmnopqrstuvwxyz` is used as an index to select a menu entry. In <https://github.com/Frenzie/nimbler> I also used uppercase should lowercase run out, but I don't believe that's necessary here.

0 closes the menu at the root or goes back one level when deeper down.

The remainder is automatically assigned a clockwise number, starting from the topmost item.

References #291. To close it, presumably some display is desired as well.